add a "dry-run" option

+++ src/main/scala/Main.scala
@@ -18,12 +18,12 @@ object Main extends App {
     the default thread pool uses daemon threads which don't hold us open until they complete.
     we want regular threads that do.
      */
-    import com.amazonaws.metrics.AwsSdkMetrics
-    AwsSdkMetrics.enableDefaultMetrics
-
-    AwsSdkMetrics.setCredentialProvider(new DefaultAWSCredentialsProviderChain())
-
-    AwsSdkMetrics.setMetricNameSpace("UploadFlushList")
+//    import com.amazonaws.metrics.AwsSdkMetrics
+//    AwsSdkMetrics.enableDefaultMetrics
+//
+//    AwsSdkMetrics.setCredentialProvider(new DefaultAWSCredentialsProviderChain())
+//
+//    AwsSdkMetrics.setMetricNameSpace("UploadFlushList")
 
     val listFileName = args.headOption match {
       case None=>
@@ -32,13 +32,27 @@ object Main extends App {
       case Some(listfile)=>listfile
     }
 
-    lazy val maxThreads = System.getProperty("maxThreads") match {
+    lazy val uploadThreads = System.getProperty("uploadThreads") match {
       case null =>
-        48
+        8
       case str: String =>
         str.toInt
     }
 
+    lazy val verifyThreads = System.getProperty("verifyThreads") match {
+      case null =>
+        36
+      case str: String =>
+        str.toInt
+    }
+
+    lazy val delayBetweenChunks = Option(System.getProperty("delayBetweenChunks")).map(_.toInt)
+
+    System.getProperty("maxThreads") match {
+      case null =>
+      case str: String =>
+        throw new RuntimeException("maxThreads is deprecated, you should specify uploadThreads and verifyThreads seperately.")
+    }
     lazy val reallyDelete = System.getProperty("reallyDelete") match {
       case null =>
         false
@@ -53,15 +67,31 @@ object Main extends App {
         Some(str.toInt)
     }
 
-    implicit val exec:ExecutionContext = ExecutionContext.fromExecutor(Executors.newFixedThreadPool((maxThreads.toFloat*2.0/3.0).toInt))
-    val uploadExecContext = ExecutionContext.fromExecutor(Executors.newFixedThreadPool((maxThreads.toFloat*1.0/3.0).toInt))
+    implicit val exec:ExecutionContext = ExecutionContext.fromExecutor(Executors.newFixedThreadPool(verifyThreads))
+    val uploadExecContext = ExecutionContext.fromExecutor(Executors.newFixedThreadPool(uploadThreads))
 
     val logger = LoggerFactory.getLogger(getClass)
     lazy val clientConfg:ClientConfiguration = new ClientConfiguration()
-    clientConfg.setMaxConnections(maxThreads)
+    clientConfg.setMaxConnections(uploadThreads)
     clientConfg.setRequestTimeout(120000)
     clientConfg.setUseReaper(true)
-    lazy implicit val s3conn: AmazonS3 = AmazonS3ClientBuilder.standard().withClientConfiguration(clientConfg).build()
+
+
+    def getS3Client:AmazonS3 = {
+      val clientBuilder = AmazonS3ClientBuilder
+        .standard()
+      val clientBuilderWithRegion = System.getProperty("region") match {
+        case null =>
+          clientBuilder
+        case regionName:String=>
+          clientBuilder.withRegion(regionName)
+      }
+      clientBuilderWithRegion
+        .withClientConfiguration(clientConfg)
+        .build()
+    }
+
+    lazy implicit val s3conn: AmazonS3 = getS3Client
 
     lazy val destBucket = System.getProperty("destBucket")
     if (destBucket == null) {
@@ -98,6 +128,18 @@ object Main extends App {
         true
     }
 
+    lazy val dryRun = System.getProperty("dryRun") match {
+      case null=>
+        false
+      case str:String=>
+        true
+    }
+
+    lazy val pathPrefix = Option(System.getProperty("pathPrefix")) match {
+      case pxf @ Some(_)=>pxf
+      case None=>Option(System.getenv("pathPrefix"))
+    }
+
     lazy val storageClass:StorageClass = System.getProperty("storageClass") match {
       case null=>StorageClass.StandardInfrequentAccess
       case str:String =>StorageClass.fromValue(str)
@@ -109,8 +151,10 @@ object Main extends App {
     logger.info(s"Removing $pathSegments from paths for upload")
     logger.info(s"Uploading to $destBucket")
     logger.info(s"Really delete is $reallyDelete")
+    logger.info(s"Upload threads: $uploadThreads, verify threads: $verifyThreads")
     logger.info(s"noProjects is set to $noProjects")
     logger.info(s"Storage class is $storageClass")
+    logger.info(s"Dryrun is $dryRun")
 
     val lp = new ListParser(listFileName, noProjects)
 
@@ -138,31 +182,38 @@ object Main extends App {
         false
       } else {
         n+=1
-        uploader.kickoff_upload(filePath, uploadExecContext).map(uploadResult => {
+        uploader.kickoff_upload(filePath, pathPrefix, dryRun, uploadExecContext, delayBetweenChunks = delayBetweenChunks).map(uploadResult => {
           if(uploadResult.uploadType==UploadResultType.AlreadyThere) alreadyUploadedCounter+=1
-          logger.info(s"$filePath: upload completed successfully (${uploadResult.uploadType.toString}), calculating etag")
-          EtagCalculator.propertiesForFile(new File(filePath), chunkSize)(exec).map(localFileProperties => {
-            logger.debug(s"$filePath: etag calculated, checking if deletable")
-            FileChecker.canDelete(destBucket, uploadResult.uploadedPath, localFileProperties) match {
-              case true =>
-                successfulCounter+=1
-                if(reallyDelete) {
-                  logger.info(s"$filePath: will delete")
-                  fileref.delete()
-                } else {
-                  logger.info(s"$filePath: can be deleted. Set reallyDelete property to 'true'")
-                }
-              case false =>
-                failedCounter+=1
-                logger.warn(s"$filePath: UPLOAD FAILED, removing remote path")
-                uploader.delete_failed_upload(filePath) match {
-                  case Success(u)=>
-                    logger.debug(s"$filePath: remote file deleted")
-                  case Failure(err)=>
-                    logger.error(s"$filePath: could not delete remote file", err)
-                }
-            }
-          })
+          val uploadVerb = uploadResult.uploadType match {
+            case UploadResultType.AlreadyThere=>"already present"
+            case UploadResultType.DryRun=>"would have been uploaded"
+            case _=>"completed successfully"
+          }
+          logger.info(s"$filePath: upload $uploadVerb (${uploadResult.uploadType.toString}) to ${uploadResult.uploadedPath}, calculating etag")
+          if(!dryRun) {
+            EtagCalculator.propertiesForFile(new File(filePath), chunkSize)(exec).map(localFileProperties => {
+              logger.debug(s"$filePath: etag calculated, checking if deletable")
+              FileChecker.canDelete(destBucket, uploadResult.uploadedPath, localFileProperties) match {
+                case true =>
+                  successfulCounter += 1
+                  if (reallyDelete) {
+                    logger.info(s"$filePath: will delete")
+                    fileref.delete()
+                  } else {
+                    logger.info(s"$filePath: can be deleted. Set reallyDelete property to 'true'")
+                  }
+                case false =>
+                  failedCounter += 1
+                  logger.warn(s"$filePath: UPLOAD FAILED, removing remote path")
+                  uploader.delete_failed_upload(filePath, pathPrefix) match {
+                    case Success(u) =>
+                      logger.debug(s"$filePath: remote file deleted")
+                    case Failure(err) =>
+                      logger.error(s"$filePath: could not delete remote file", err)
+                  }
+              }
+            })
+          }
           if(limit.isDefined){
             if(n>limit.get){
               logger.info(s"Already triggered $n items, stopping")
